% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funciones_habits.R
\name{cual_id}
\alias{cual_id}
\title{Que ID corresponde usar para pegar habits}
\usage{
cual_id(df)
}
\arguments{
\item{df}{dataframe con las columnas ID_COM, ID_DIS e ID_MAN obtenidas a partir de la geolocalización y los mapas del país}
}
\description{
A partir de un dataframe que contiene las columnas ID_COM, ID_DIS e ID_MAN devuelve una columna con la ID que corresponde usar para pegar habits
}
\examples{

}
\keyword{habits}
\keyword{id}
