% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funciones_portal.R
\name{dist_imap}
\alias{dist_imap}
\title{dist_imap}
\usage{
dist_imap(df_f, registro)
}
\arguments{
\item{df_f}{dataframes de comparables. Debe tener columnas lat y lng sin NAs}

\item{registro}{propiedad para la cual se buscan testigos}
}
\description{
Encuentra las distancias a los comparables (propiedades similares) de una propiedad en particular a la base de testigos
}
\keyword{distancias}
\keyword{levershine}
\keyword{propiedades}
\keyword{testigos}
