% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funciones_habits.R
\name{elegir_id}
\alias{elegir_id}
\title{Elegir ID que finalmente se va a utilizar}
\usage{
elegir_id(df)
}
\arguments{
\item{df}{dataframe con las columnas cual_id_usar, ID_COM, ID_DIS e ID_MAN obtenidas a partir de la geolocalización y los mapas del país}
}
\description{
A partir de un dataframe que contiene las columnas cual_id_usar, ID_COM, ID_DIS e ID_MAN selecciona la ID para hacer mas facil un siguiente merge con habits
}
\examples{

}
\keyword{habits}
\keyword{id}
