% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funciones_entorno.R
\name{entorno_farmacias}
\alias{entorno_farmacias}
\title{entorno_farmacias}
\usage{
entorno_farmacias(registro, base_entorno, distancia = 500)
}
\arguments{
\item{registro}{propiedad geolocalizada para la que se buscan las variables de colegios. Deberia ser un data frame de una fila con columnas lat y lng}

\item{base_entorno}{base de datos de los colegios geolocalizados}

\item{distancia}{distancia a buscar farmacias. Default 500 para "cerca", distancia media se toma el doble, default 1000}
}
\description{
Crea variables relacionada a la distancia de la propiedad a la base geolocalizada de farmacias. Variables: se encuentra farmacia cerca, se encuentra farmacia a distancia media, cantidad de farmacias a distancia media
}
\keyword{entorno}
\keyword{farmacias}
\keyword{variables}
