% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funciones.R, R/funciones_geoloc.R
\name{geocode_it}
\alias{geocode_it}
\title{geocode_it}
\usage{
geocode_it(direccion, key = NULL)

geocode_it(direccion, key = NULL)
}
\arguments{
\item{direccion}{direccion de la propiedad. Idealmente Calle + Numero + Comuna + Pais}

\item{key}{identificación para poder utilizar el servicio de google. Para mas información ver como obtener key en google. Actualmente hay una asociada a cuenta aischile.pruebas@gmail.com, pass aischile902}

\item{direccion}{direccion de la propiedad. Idealmente Calle + Numero + Comuna + Pais}

\item{key}{identificación para poder utilizar el servicio de google. Para mas información ver como obtener key en google. Actualmente hay una asociada a cuenta aischile.pruebas@gmail.com, pass aischile902}
}
\description{
Geolocaliza una dirección (texto) devolviendo un data frame con latitud y longitud

Geolocaliza una dirección (texto) devolviendo un data frame con latitud y longitud
}
\examples{


}
\keyword{direccion}
\keyword{geolocalizar}
