% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funciones_habits.R
\name{get_id}
\alias{get_id}
\title{Encuentra ID del distrito para un dataframe de coordenadas}
\usage{
get_id(coordenadas, mapa, nivel = "COD_DISTRI")
}
\arguments{
\item{coordenadas}{dataframe con columnas lat y lng numéricas sin NAs}

\item{mapa}{objeto SpatialPolygonsDataFrame (i.e un .shp leido en R) para la región de las coordenadas}

\item{nivel}{a que nivel se busca la ID ('COD_DISTRI', 'COMUNA' o 'MANZENT' para distrito, comuna y manzana respectivamente para mapas usados en WebService 2018). Default 'COD_DISTRI'}
}
\description{
A partir de un dataframe de lng/lat (en ese orden y sin NA) encuentra la ID del distrito al que corresponden las coordenadas
}
\examples{

}
\keyword{distrito}
\keyword{habits}
\keyword{id}
