% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funciones.R, R/funciones_geoloc.R
\name{norm1}
\alias{norm1}
\title{norm1}
\usage{
norm1(direcciones)

norm1(direcciones)
}
\arguments{
\item{direcciones}{direccion (caracter) a geolocalizar}

\item{direcciones}{direccion (caracter) a geolocalizar}
}
\description{
Estandariza una direccion arreglando errores comunes y exceso de informacion

Estandariza una direccion arreglando errores comunes y exceso de informacion
}
\examples{
norm1("Pje. Dominic Dicoco 15 bod 15 dp 34 r ")
norm1("Pje. Dominic Dicoco 15 bod 15 dp 34 r ")
}
\keyword{direccion}
\keyword{geolocalizar}
