% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funciones.R, R/funciones_geoloc.R
\name{norm2}
\alias{norm2}
\title{norm2}
\usage{
norm2(direccion)

norm2(direccion)
}
\arguments{
\item{direcciones}{direccion (caracter) a geolocalizar}

\item{direcciones}{direccion (caracter) a geolocalizar}
}
\description{
Se queda solo con la parte de la direccion relevante DIRECCION + NUMERO

Se queda solo con la parte de la direccion relevante DIRECCION + NUMERO
}
\examples{
norm2("CALLE SECRETA 123 DP 1293 BOD 45")
norm2("CALLE SECRETA 123 DP 1293 BOD 45")
}
\keyword{direccion}
\keyword{geolocalizar}
