% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funciones_modelo.R
\name{predecir}
\alias{predecir}
\title{dist_imap}
\usage{
predecir(registro, varmod, cols, xgbcv)
}
\arguments{
\item{registro}{propiedad para la cual se buscan testigos}

\item{df_f}{dataframes de comparables. Debe tener columnas lat y lng sin NAs}
}
\description{
Encuentra las distancias a los comparables (propiedades similares) de una propiedad en particular a la base de testigos
}
\keyword{distancias}
\keyword{levershine}
\keyword{propiedades}
\keyword{testigos}
