% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funciones.R, R/funciones_portal.R
\name{reducir_base_comparar}
\alias{reducir_base_comparar}
\title{reducir_base_comparar}
\usage{
reducir_base_comparar(registro, df, distancia = 500,
  variable_metraje = "sum_V8_b", similitud = 0.2, n_testigos = 4)

reducir_base_comparar(registro, df, distancia = 500,
  variable_metraje = "sum_V8_b", similitud = 0.2, n_testigos = 4)
}
\arguments{
\item{registro}{propiedad para la cual se quiere encontrar comparables. Deberia ser un data frame de una fila con la variable sol.sol_tip_bie, ("URCS CASAS"), cod_comuna_habits y cod_region_habits}

\item{df}{propiedad para la cual se quiere encontrar comparables. Deberia ser un data frame de una fila con la variable sol.sol_tip_bie, ("URCS CASAS")}

\item{distancia}{distancia maxima a propiedades comparables. Default 500 metros}

\item{variable_metraje}{el nombre de la variable de metraje en registro. Default sum_V8_b porque ese es el nombre del metraje en base SII}

\item{similitud}{valor entre 0-1 para comparar propiedades. Default 0.2 lo que significa que busca comparables que sean +-20% del metraje del registro}

\item{disntancia}{distancia maxima a propiedades comparables. Default 500 metros}

\item{registro}{propiedad para la cual se quiere encontrar comparables. Deberia ser un data frame de una fila con la variable sol.sol_tip_bie, ("URCS CASAS"), cod_comuna_habits y cod_region_habits}

\item{df}{propiedad para la cual se quiere encontrar comparables. Deberia ser un data frame de una fila con la variable sol.sol_tip_bie, ("URCS CASAS")}
}
\description{
Encuentra las distancias a los comparables (propiedades similares) de una propiedad en particular a la base de testigos

Encuentra las distancias a los comparables (propiedades similares) de una propiedad en particular a la base de testigos
}
\examples{
norm2("CALLE SECRETA 123 DP 1293 BOD 45")
}
\keyword{direccion}
\keyword{geolocalizar}
