% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funciones_portal.R
\name{valores}
\alias{valores}
\title{valores}
\usage{
valores(df)
}
\arguments{
\item{df}{dataframes de testigos}
}
\description{
Calcula variables a partir de base de testigos. Precio promedio, precio m2 promedio, metraje promedio y variable secreta
}
\keyword{distancias}
\keyword{levershine}
\keyword{propiedades}
\keyword{testigos}
