% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funciones_portal.R
\name{valores_aux}
\alias{valores_aux}
\title{valores_aux}
\usage{
valores_aux(df)
}
\arguments{
\item{df}{dataframes de comparables. Debe tener columnas lat y lng sin NAs}
}
\description{
Encuentra las distancias a los comparables (propiedades similares) de una propiedad en particular a la base de testigos
}
\keyword{distancias}
\keyword{levershine}
\keyword{propiedades}
\keyword{testigos}
