% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funciones_modelo.R
\name{predecir}
\alias{predecir}
\title{dist_imap}
\usage{
predecir(registro, varmod, vars_onehot, modelo)
}
\arguments{
\item{registro}{propiedad para la cual se buscan testigos}

\item{varmod}{nombres de variables que el modelo necesita}

\item{vars_onehot}{nombres de variables onehot}

\item{modelo}{nombres de variables onehot}
}
\description{
Encuentra las distancias a los comparables (propiedades similares) de una propiedad en particular a la base de testigos
}
\keyword{distancias}
\keyword{levershine}
\keyword{propiedades}
\keyword{testigos}
